Name the conversions that lose something, and deny `as`
A domain wrapper proves a `Count` is not a `Length`, and a `Ratio` is in
0..=1 because its constructor refused everything else. It cannot prove a
runtime `u64` fits an `i64` — that is a statement about a value, and Rust
has no refinement types for it. So at every boundary to a foreign API the
wrapper ends and a conversion begins, and those were spelled `as`, the one
cast that wraps silently.

`From` is the wrong home for it: `From`/`Into` promise the value survives,
which is why std ships `TryFrom<u64> for i64` and no `From`. It would also
hand out `.into()`, so the loss would land at call sites that never named
it. And a blanket `unwrap_or(MAX)` is wrong for a signed source —
`u32::try_from(-5i64)` would give 4294967295 instead of 0.

Two traits instead, each named for what it does, so the three things `as`
means are told apart where they happen:

    exact              From / Into (std)
    out of range       SaturatingInto — stops at the nearer end
    not representable  ApproxInto — the nearest number that fits

...and then `as` itself is denied, in a `[workspace.lints.clippy]` all three
crates inherit. Five lints, and the last is the point: `cast_lossless` is
what makes an exact conversion say `From` rather than reach for one of the
other two.

The traits and the denial are one change, which is why they are one commit.
Without the denial nothing stops the next `as`, and `ApproxInto` in
particular changes no behaviour on its own — the float directions of `as`
already do the right thing (RFC 2484). Its job is to keep a denied cast from
being holed by an `#[allow]`: a line in perks.rs carries an int-to-float
cast next to a float-to-int one, and with no name for the first, the only
way to compile it would be an allow covering both.

Only int-to-int changes what the program does; there `as` wraps and the
trait clamps. Two call sites had already grown their own clamp by hand —
`value.min(i64::MAX as u64) as i64` and `clamp(0, u32::MAX as i64) as u32` —
and now say it in one word. Where the sink is ours, `Gauge::set` takes the
trait, so its callers hand over the domain value itself.

The macro's generated division went through `as` too, where the denial
cannot reach: a proc macro's spans are exempt. It names its conversion now,
which does not fix the precision the TODO there describes but stops it being
invisible.

Three casts survive, none of them flagged: `group as u8` is an enum
discriminant, `chat_id.0 as i64` is sqlx's type-override syntax rather than
a cast, and `UID as u64` sits inside `literal!`, so the build evaluates it.

Closes #161.

+++ src/scheduler/deletions.rs
@@ -49,7 +49,7 @@ pub async fn run_pending_deletions(
     let due = repos.deletions.claim_due(config.batch_size, Utc::now() + config.lease).await?;
     // The empty runs are measured too: an idle worker is what tells a queue that keeps up from one
     // that is merely being asked for less than it holds.
-    metrics::SELF_DESTRUCTION_BATCH_SIZE.observe(due.len() as f64);
+    metrics::SELF_DESTRUCTION_BATCH_SIZE.observe(due.len());
     if due.is_empty() {
         return Ok(())
     }
